🐞 fixdeps: add `getDeps` function
