+++ packages/fixdeps/src/index.ts
@@ -1,2 +1,3 @@
+export * from './get-deps'
 export * from './fixdeps'
 export * from './types'
