+++ packages/fixdeps/src/types.ts
@@ -28,6 +28,13 @@ export type TPackageJson = {
   },
 } & TJsonMap
 
+export type TGetDepsResult = {
+  depsToAdd: string[],
+  depsToRemove: string[],
+  devDepsToAdd: string[],
+  peerDevDepsToAdd: string[],
+}
+
 export type TResult = {
   addedDeps: TDepsObject,
   addedDevDeps: TDepsObject,
